% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{fullselection}
\alias{fullselection}
\title{Full Selection Pipeline Wrapper}
\usage{
fullselection()
}
\value{
A function that applies the pipeline steps to a \code{varselectPipeline} object.
}
\description{
A convenience wrapper that runs the complete default pipeline: Validate -> Significance -> Refine -> Selection -> Prediction.
}
