% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{+,varselectPipeline,function-method}
\alias{+,varselectPipeline,function-method}
\title{Chain Pipeline Steps}
\usage{
\S4method{+}{varselectPipeline,function}(e1, e2)
}
\arguments{
\item{e1}{A \code{varselectPipeline} object.}

\item{e2}{A function returning a modified \code{varselectPipeline} object.}
}
\value{
A modified \code{varselectPipeline} object.
}
\description{
Enables the use of the `+` operator to chain pipeline functions.
}
