% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{prediction}
\alias{prediction}
\title{Assess Prediction}
\usage{
prediction(metric = "rmse", cv_folds = 10)
}
\arguments{
\item{metric}{Primary metric to display (default "rmse").}

\item{cv_folds}{Number of CV folds for validation.}
}
\description{
Calculates final performance metrics (RMSE, MAE, R2 for regression; AUC for classification) using Cross-Validation.
}
