% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{refine}
\alias{refine}
\title{Refine by Multicollinearity}
\usage{
refine(vif_threshold = 10, condition_threshold = 30)
}
\arguments{
\item{vif_threshold}{VIF threshold to trigger removal (default 10).}

\item{condition_threshold}{Condition number threshold (default 30).}
}
\description{
Removes variables using VIF and Condition Number analysis to reduce multicollinearity.
}
