% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{validate}
\alias{validate}
\title{Validate Data}
\usage{
validate(exclude = NULL, missing_threshold = 0.3, variance_threshold = 0.01)
}
\arguments{
\item{exclude}{Vector of variable names to exclude manually.}

\item{missing_threshold}{Max allowed proportion of missing values (default 0.3).}

\item{variance_threshold}{Min required variance (default 0.01).}
}
\value{
A closure that modifies the pipeline object.
}
\description{
Removes variables based on missingness, variance, or user exclusion.
}
