% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{varselect}
\alias{varselect}
\title{Variable Selection Pipeline Package}
\usage{
varselect(data, target)
}
\arguments{
\item{data}{A data frame containing target and predictors.}

\item{target}{A character string specifying the target variable name.}
}
\value{
An object of class \code{varselectPipeline}.
}
\description{
A pipeline-based approach to variable selection in R using method chaining.

Creates a new variable selection pipeline object.
}
