% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\docType{class}
\name{varselectPipeline-class}
\alias{varselectPipeline-class}
\title{Pipeline Class for Variable Selection}
\description{
S4 class to store the state of the variable selection pipeline.
}
\section{Slots}{

\describe{
\item{\code{data}}{The input data frame.}

\item{\code{target}}{The name of the target variable.}

\item{\code{predictors}}{Vector of predictor names.}

\item{\code{target_type}}{Type of target ("numeric", "binary", "multiclass").}

\item{\code{selected_vars}}{Variables selected during the selection phase.}

\item{\code{final_vars}}{Final subset of variables used for the model.}

\item{\code{scores}}{Statistical scores for variables.}

\item{\code{performance}}{List of model performance metrics.}

\item{\code{model}}{The trained model object.}

\item{\code{steps}}{History of steps applied.}

\item{\code{full_mse}}{MSE of the full model (used for Mallow's Cp).}
}}

