% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{visualize}
\alias{visualize}
\title{Visualize Importance}
\usage{
visualize(type = "importance", top = 5)
}
\arguments{
\item{type}{Type of plot (default "importance").}

\item{top}{Number of top variables to show.}
}
\description{
Plots variable importance based on final model coefficients.
}
